requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.0
aiocron>=1.8
pydantic>=2.7
langchain-community>=0.2.5
langchain-chroma>=0.1.2
langchain-huggingface>=0.0.6
langchain-ollama>=0.1.0
langchain-text-splitters>=0.0.3
langchain-unstructured>=0.0.2
python-multipart
sentence-transformers
unstructured
pandas
python-docx
msoffcrypto-tool
streamlit
PyJWT
bcrypt